pyproject.toml
[project]
name = "langchain-curso-alura"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]
